requirements.txt
azure-cognitiveservices-speech==1.32.1
fastapi==0.103.0
pydantic==1.10.12
pydub==0.25.1
python-dotenv==0.21.0
pytube==15.0.0
redis==5.0.0
rq==1.15.1
shortuuid==1.0.11
uvicorn==0.18.3
psycopg2-binary==2.9.7